chore: refactor container utils out into test_utils.
