+++ gwrdfa/resample/core/src/agreement/data.rs
@@ -78,193 +78,4 @@ pub mod test {
 	use parabyzantine::agreement::{
 		Agreement, ParabyzantineAgreementData, ParabyzantineAgreementDataSpec,
 	};
-
-	/// A wrapper for an index s.t. we guarantee disjoint types.
-	#[derive(Debug, Clone, Default, PartialEq, Eq)]
-	pub struct Index<T: Eq>(pub T);
-
-	/// A wrapper for a value s.t. we guarantee disjoint types.
-	#[derive(Debug, Clone, Default, PartialEq, Eq)]
-	pub struct Value<T: Eq + 'static>(pub T);
-
-	/// A wrapper for a subcommittee s.t. we guarantee disjoint types.
-	#[derive(Debug, Clone, Default, PartialEq, Eq)]
-	pub struct Sub<T: Eq>(pub T);
-
-	/// A container for a certificate.
-	pub struct TestResampleCertificateContainer<I: Eq, V: Eq + 'static, S: Subcommittee<V>> {
-		pub index: Component<Index<I>>,
-		pub value: Component<Value<V>>,
-		pub subcommittee: Component<Sub<S>>,
-	}
-
-	/// A [ContainerGiving] implementation for [Index<I>].
-	impl<I: Eq, V: Eq + 'static, S: Subcommittee<V>> ContainerGiving<Index<I>>
-		for TestResampleCertificateContainer<I, V, S>
-	{
-		fn as_component(&self) -> Component<&Index<I>> {
-			self.index.as_ref()
-		}
-	}
-
-	/// A [ContainerGiving] implementation for [Value<V>].
-	impl<I: Eq, V: Eq + 'static, S: Subcommittee<V>> ContainerGiving<Value<V>>
-		for TestResampleCertificateContainer<I, V, S>
-	{
-		fn as_component(&self) -> Component<&Value<V>> {
-			self.value.as_ref()
-		}
-	}
-
-	/// A [ContainerGiving] implementation for [Sub<S>].
-	impl<I: Eq, V: Eq + 'static, S: Subcommittee<V>> ContainerGiving<Sub<S>>
-		for TestResampleCertificateContainer<I, V, S>
-	{
-		fn as_component(&self) -> Component<&Sub<S>> {
-			self.subcommittee.as_ref()
-		}
-	}
-
-	/// We want the delta buffer to be able to hold the deltas for the index, value, and subcommittee.
-	pub struct TestResampleCertificateDelta<I: Eq, V: Eq + 'static, S: Subcommittee<V>> {
-		pub index: Delta<Index<I>>,
-		pub value: Delta<Value<V>>,
-		pub subcommittee: Delta<Sub<S>>,
-	}
-
-	/// A [DeltasContainer] implementation for [TestResampleCertificateContainer<I, V, S>].
-	impl<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>>
-		DeltasContainer<TestResampleCertificateContainer<Index, Value, Sub>>
-		for TestResampleCertificateDelta<Index, Value, Sub>
-	{
-		fn apply_deltas(self, container: &mut TestResampleCertificateContainer<Index, Value, Sub>) {
-			self.index.apply(&mut container.index);
-			self.value.apply(&mut container.value);
-			self.subcommittee.apply(&mut container.subcommittee);
-		}
-
-		fn into_container(self) -> TestResampleCertificateContainer<Index, Value, Sub> {
-			TestResampleCertificateContainer {
-				index: self.index.into_component(),
-				value: self.value.into_component(),
-				subcommittee: self.subcommittee.into_component(),
-			}
-		}
-	}
-
-	/// A container for an agreement.
-	pub struct TestResampleAgreementContainer<
-		Index: Eq,
-		Value: Eq + 'static,
-		Sub: Subcommittee<Value>,
-	> {
-		pub agreement: Component<Agreement>,
-		pub resample: Component<Resample>,
-		pub index: Component<Index>,
-		pub value: Component<Value>,
-		pub subcommittee: Component<Sub>,
-	}
-
-	/// A [DeltasContainer] implementation for [TestResampleAgreementContainer<I, V, S>].
-	pub struct TestResampleAgreementDelta<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>> {
-		pub agreement: Delta<Agreement>,
-		pub resample: Delta<Resample>,
-		pub index: Delta<Index>,
-		pub value: Delta<Value>,
-		pub subcommittee: Delta<Sub>,
-	}
-
-	/// A [DeltasContainer] implementation for [TestResampleAgreementContainer<I, V, S>].
-	impl<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>>
-		DeltasContainer<TestResampleAgreementContainer<Index, Value, Sub>>
-		for TestResampleAgreementDelta<Index, Value, Sub>
-	{
-		fn apply_deltas(self, container: &mut TestResampleAgreementContainer<Index, Value, Sub>) {
-			self.agreement.apply(&mut container.agreement);
-			self.resample.apply(&mut container.resample);
-			self.index.apply(&mut container.index);
-			self.value.apply(&mut container.value);
-			self.subcommittee.apply(&mut container.subcommittee);
-		}
-
-		fn into_container(self) -> TestResampleAgreementContainer<Index, Value, Sub> {
-			TestResampleAgreementContainer {
-				agreement: self.agreement.into_component(),
-				resample: self.resample.into_component(),
-				index: self.index.into_component(),
-				value: self.value.into_component(),
-				subcommittee: self.subcommittee.into_component(),
-			}
-		}
-	}
-
-	pub struct TestResampleParabyzantineSpec<
-		Index: Eq,
-		Value: Eq + 'static,
-		Sub: Subcommittee<Value>,
-	> {
-		__marker: PhantomData<(Index, Value, Sub)>,
-	}
-
-	impl<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>> ParabyzantineAgreementDataSpec
-		for TestResampleParabyzantineSpec<Index, Value, Sub>
-	{
-		type CertificateEntity = ContainerEntity;
-		type CertificateBuffer =
-			ContainerEntityBuffer<TestResampleCertificateContainer<Index, Value, Sub>>;
-		type CertificateDraftBuffer =
-			ContainerEntityDraftBuffer<TestResampleCertificateDelta<Index, Value, Sub>>;
-		type AgreementEntity = ContainerEntity;
-		type AgreementBuffer =
-			ContainerEntityBuffer<TestResampleAgreementContainer<Index, Value, Sub>>;
-		type AgreementDraftBuffer =
-			ContainerEntityDraftBuffer<TestResampleAgreementDelta<Index, Value, Sub>>;
-	}
-
-	pub struct TestResampleParabyzantineData<
-		Index: Eq,
-		Value: Eq + 'static,
-		Sub: Subcommittee<Value>,
-	> {
-		pub certificate_buffer:
-			ContainerEntityBuffer<TestResampleCertificateContainer<Index, Value, Sub>>,
-		pub agreement_buffer:
-			ContainerEntityBuffer<TestResampleAgreementContainer<Index, Value, Sub>>,
-	}
-
-	impl<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>>
-		ParabyzantineAgreementData<TestResampleParabyzantineSpec<Index, Value, Sub>>
-		for TestResampleParabyzantineData<Index, Value, Sub>
-	{
-		fn parabyzantine_agreement_certificate_buffer(
-			&self,
-		) -> &ContainerEntityBuffer<TestResampleCertificateContainer<Index, Value, Sub>> {
-			&self.certificate_buffer
-		}
-		fn parabyzantine_agreement_certificate_buffer_mut(
-			&mut self,
-		) -> &mut ContainerEntityBuffer<TestResampleCertificateContainer<Index, Value, Sub>> {
-			&mut self.certificate_buffer
-		}
-		fn parabyzantine_agreement_certificate_draft_buffer(
-			&self,
-		) -> ContainerEntityDraftBuffer<TestResampleCertificateDelta<Index, Value, Sub>> {
-			ContainerEntityDraftBuffer::new()
-		}
-		fn parabyzantine_agreement_agreement_buffer(
-			&self,
-		) -> &ContainerEntityBuffer<TestResampleAgreementContainer<Index, Value, Sub>> {
-			&self.agreement_buffer
-		}
-		fn parabyzantine_agreement_agreement_buffer_mut(
-			&mut self,
-		) -> &mut ContainerEntityBuffer<TestResampleAgreementContainer<Index, Value, Sub>> {
-			&mut self.agreement_buffer
-		}
-		fn parabyzantine_agreement_agreement_draft_buffer(
-			&self,
-		) -> ContainerEntityDraftBuffer<TestResampleAgreementDelta<Index, Value, Sub>> {
-			ContainerEntityDraftBuffer::new()
-		}
-	}
 }
