+++ gwrdfa/resample/core/src/agreement/test_util/container.rs
@@ -3,3 +3,9 @@ pub use agreement::*;
 
 pub mod certificate;
 pub use certificate::*;
+
+pub mod spec;
+pub use spec::*;
+
+pub mod data;
+pub use data::*;
